requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
SQLAlchemy==1.4.53
pydantic==1.10.11
qrcode==7.3
Pillow==10.0.0
PyJWT==2.8.0